task: Rename testing directories
