+++ tests/get-severity-test.ts
