+++ vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
   test: {
-    include: ['__tests__/**/*-test.ts'],
+    include: ['tests/**/*-test.ts'],
     global: true,
   },
 });
